+++ src/Project/shapes/Bezier.cpp
@@ -8,16 +8,18 @@
  *  @version 1.0.0
  *  @license MIT (https://mit-license.org)
  */
+// TODO: Test Bezier curve
 #include "Bezier.h"
 #include "math.h"
 
 /**
- * Create a new bezier curve
- * @param p0    POS   start point
- * @param p1    POS   1st control point
- * @param p2    POS   2nd control point
- * @param p3    POS   end point
- * @param drive Drive drive controller
+ * Bezier curve
+ * @param p0    Start point
+ * @param p1    First control point
+ * @param p2    Second control point
+ * @param p3    End point
+ * @param drive Drive controller
+ * @param lcd   LCD screen controller
  */
 Bezier::Bezier(POS p0, POS p1, POS p2, POS p3, Drive *drive, LiquidCrystal *lcd):
     Shape(drive, lcd),
@@ -27,10 +29,9 @@ Bezier::Bezier(POS p0, POS p1, POS p2, POS p3, Drive *drive, LiquidCrystal *lcd)
     _p3(p3){};
 
 /**
- * Draw the bezier curve
- * @param  p bool print info to client
- * @return   POS  the current position
- *
+ * Draw the Bezier curve
+ * @param  p Whether or not to print details
+ * @return   Updated position
  Latex:
 $$
 x_t = x_0 + 3t(x_1 - x_0) + 3t^2(x_0 + x_2 - 2x_1) + t^3(x_3 - x_0 + 3x_1 - 3x_2) \\
@@ -42,46 +43,56 @@ POS Bezier::draw(bool p) {
 
     if(p) print();
 
+    // Move to start point
     _drive->moveTo(_p0.x, _p0.y);
 
-    if(p) Serial.println("L");
+    // TODO: Update resolution to get a proper total number of x steps
 
+    // Get our smallest x
     int min = _p0.x;
     if(_p1.x < min) min = _p1.x;
     if(_p2.x < min) min = _p2.x;
     if(_p3.x < min) min = _p3.x;
 
+    // Get our largest x
     int max = _p0.x;
     if(_p1.x > max) max = _p1.x;
     if(_p2.x > max) max = _p2.x;
     if(_p3.x > max) max = _p3.x;
 
+    // (max - min) = t steps (resolution)
     for(int i=0; i<=(max - min); i++){
-        double t = i/(max - min);
-        double t2 = pow(t, 2);
-        double t3 = pow(t, 3);
+        double t = i/(max - min); // 0 <= t <= 1
+        double t2 = pow(t, 2);    // t^2
+        double t3 = pow(t, 3);    // t^3
 
+        // Get our x value
         int x = (int)(
             _p0.x +
             3 * t * (_p1.x - _p0.x) +
             3 * t2 * (_p0.x + _p2.x - 2 * _p1.x) +
             t3 * (_p3.x - _p0.x + 3 * _p1.x - 3 * _p2.x)
         );
+
+        // Get our y value
         int y = (int)(
             _p0.y +
             3 * t * (_p1.y - _p0.y) +
             3 * t2 * (_p0.y + _p2.y - 2 * _p1.y) +
             t3 * (_p3.y - _p0.y + 3 * _p1.y - 3 * _p2.y)
         );
+
+        // Draw to our next value
         _drive->lineTo(x, y);
     }
 
+    // Return updated position
     return _drive->get();
 };
 
 /**
- * Print the bezier info to the client
+ * Print details to LCD and Serials
  */
 void Bezier::print() {
-
+    // TODO: write up print function for Bezier curve
 };