+++ src/Project/shapes/Polygon.h
@@ -18,15 +18,38 @@
  */
 class Polygon: public Shape {
 private:
-    LinkedList<POS> *_points; // Points to draw between
+    LinkedList<POS> *_points; // Points to draw between, first point is moveTo
 
+    /**
+     * Print details to lcd and Serial
+     */
     void print();
 
 public:
-    Polygon(){}; // default constructor
-    Polygon(LinkedList<POS> *points, Drive *drive, LiquidCrystal *lcd); // primary constructor
+    /**
+     * Polygon()
+     */
+    Polygon(){};
 
-    POS draw(bool p); // draw function
+    /**
+     * Polygon with a list of points to draw
+     * @param points List<POS> points to draw
+     * @param drive  Drive controller
+     * @param lcd    LCD screen controller
+     */
+    Polygon(LinkedList<POS> *points, Drive *drive, LiquidCrystal *lcd);
+
+    /**
+     * Draw from point to point
+     * @param  p Print details
+     * @return   Updated position
+     */
+    POS draw(bool p);
+
+    /**
+     * Draw without printing details
+     * @return Updated position
+     */
     POS draw(){ return draw(false); };
 };