+++ src/Project/shapes/Shape.h
@@ -5,7 +5,7 @@
  *  to call children Shape::draw().
  *
  *  @author Drew Sommer
- *  @version 1.0.0
+ *  @version 1.0.1
  *  @license MIT (https://mit-license.org)
  */
 #ifndef SHAPE_H
@@ -24,6 +24,8 @@ private:
 
     // Print function for informing the client of what is being drawn
     void print(){
+        _lcd->setCursor(0,1);
+        _lcd->print("Shape()");
         Serial.println("Shape()");
     };
 
@@ -32,12 +34,38 @@ public:
     Drive *_drive; // Drive controller, accessible by subclasses
     LiquidCrystal *_lcd;
 
-    Shape(){}; // Default constructor
-    Shape(Drive *drive, LiquidCrystal *lcd); // Normal constructor
-    ~Shape(); // Destructor
+    /**
+     * Shape()
+     */
+    Shape(){};
 
-    virtual POS draw(bool p); // Default draw function
+    /**
+     * Shape with passed parameters
+     * @param drive Drive controller
+     * @param lcd   LCD screen controller
+     */
+    Shape(Drive *drive, LiquidCrystal *lcd);
+
+    /**
+     * Destructor
+     */
+    ~Shape(){};
+
+    /**
+     * Draw our shape
+     * @param  p Boolean, whether or not to print our details
+     * @return   Updated position
+     */
+    virtual POS draw(bool p);
+
+    /**
+     * Draw our shape (without printing details)
+     * @return Updated position
+     */
     virtual POS draw(){ return draw(false); };
+
+    // Virtual allows for polymorphism and creating an array of shapes (Ellipses,
+    // Circles, Etc.) and call their respective draw functions.
 };
 
 #endif