+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 # 1.0.0
 
 ### Upload first version
+
+# 1.1.0
+
+### Fixed lineTo() not doing proper angles
+### Added LCD feedback
+### Added system for adjusting Pen low point
+### Added button to confirm Pen position before drawing
+### Added/Improved commenting